update person class
