+++ person.rb
@@ -1,10 +1,10 @@
-require_relative './nameable'
+require './nameable'
 
 class Person < Nameable
-  attr_accessor :name, :age
   attr_reader :id, :rentals
+  attr_accessor :name, :age
 
-  def initialize(age, name = 'Unknown', parent_permission: true)
+  def initialize(age:, name: 'Unknown', parent_permission: true)
     super()
     @id = Random.rand(1..1000)
     @name = name
@@ -13,61 +13,19 @@ class Person < Nameable
     @rentals = []
   end
 
-  def add_rental(date, book)
-    Rental.new(date, book, self)
+  def add_rental(rental)
+    @rentals << rental
+  end
+
+  def can_use_service?
+    @parent_permission || of_age?
   end
 
   def correct_name
     @name
-  end
-
-  def can_use_services?
-    @parent_permission || of_age?
-  end
-
+  end  
   private
-
   def of_age?
     @age >= 18
   end
 end
-
-class Decorator < Nameable
-  attr_accessor :nameable
-
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    if @nameable.correct_name.length > 10
-      str = []
-      10.times do |i|
-        str.push(@nameable.correct_name.chars[i])
-      end
-      str.join
-    else
-      @nameable.correct_name
-    end
-  end
-end
-
-person = Person.new(22, 'maximilianus')
-person.correct_name
-capitalized_person = CapitalizeDecorator.new(person)
-capitalized_person.correct_name
-capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
-capitalized_trimmed_person.correct_name
